Firmware update fails when Windows 10 username has a space
When using nanoff for a STM32 development board (specifically
STM32F429I-DISCOVERY) the firmware update fails. This is because
the command line argument for the utility ST-LINK_CLI.exe that
contains the path to the file to be programmed has a space in
it and it is not quoted. These files have been automatically
downloaded and put in a directory under the users folder.
Quote the filename for arguments to ST-LINK_CLI.exe

Also fixed a variable typo at the same time.

Signed-off-by: Nick Schicht <nick.schicht+github@gmail.com>